moving some stuff around to attempt to simplify directory structure
