+++ qiskit/providers/aer/pulse/controllers/digest_pulse_qobj.py
@@ -21,7 +21,48 @@ from collections import OrderedDict
 import numpy as np
 from qiskit.providers.aer.aererror import AerError
 from ..pulse0.cy.utils import oplist_to_array
-from .digested_pulse_qobj import DigestedPulseQobj
+
+
+class DigestedPulseQobj:
+
+    def __init__(self):
+
+        # ####################################
+        # Some "Simulation description"
+        # ####################################
+
+        # stuff related to memory/measurements
+        self.shots = None
+        self.meas_level = None
+        self.meas_return = None
+        self.memory_slots = None
+        self.memory = None
+        self.n_registers = None
+
+        # ####################################
+        # Some Signal portion
+        # ####################################
+
+        # I think these should ultimately construct and return "Signal" objects
+
+        # specific data struct being used
+        self.pulse_array = None
+        self.pulse_indices = None
+        self.pulse_dict = None
+
+        self.qubit_lo_freq = None
+
+
+        # #############################################
+        # Mix of both signal and simulation description
+        # #############################################
+
+        # These should be turned into an internal "simulation events"
+        # structure
+
+        # "experiments" contains a combination of signal information and
+        # other experiment descriptions, which should be separated
+        self.experiments = None
 
 def digest_pulse_qobj(qobj, channels, dt, qubit_list):
     """
