+++ qiskit/providers/aer/pulse/system_models/hamiltonian_model.py
@@ -18,7 +18,7 @@
 from collections import OrderedDict
 import numpy as np
 import numpy.linalg as la
-from ..string_model_parser.string_model_parser import HamiltonianParser
+from .string_model_parser.string_model_parser import HamiltonianParser
 from qiskit.providers.aer.aererror import AerError
