+++ qiskit/providers/aer/pulse/system_models/string_model_parser/apply_str_func_to_qobj.py
