+++ qiskit/providers/aer/pulse/system_models/string_model_parser/string_model_parser.py
